Moved HttpUrlConnectionResponse into sub-package 'implementation'. Removed Javadoc fragment
