+++ sdk/core/azure-core-http-httpurlconnection/src/main/java/com/azure/core/http/httpurlconnection/HttpUrlConnectionClient.java
@@ -1,6 +1,7 @@
 package com.azure.core.http.httpurlconnection;
 
 import com.azure.core.http.*;
+import com.azure.core.http.httpurlconnection.implementation.HttpUrlConnectionResponse;
 import com.azure.core.util.Context;
 import reactor.core.publisher.Mono;
